🎨 Split notifications in files per event
