+++ packages/applications/notifications/src/subscribers/utilisateur/utilisateurInvité.notification.ts
@@ -1,0 +1,44 @@
+import { match, P } from 'ts-pattern';
+
+import { Routes } from '@potentiel-applications/routes';
+import { UtilisateurInvitéEvent } from '@potentiel-domain/utilisateur';
+
+export async function utilisateurInvitéNotification({
+  payload: { identifiantUtilisateur, rôle },
+}: UtilisateurInvitéEvent) {
+  const { BASE_URL } = process.env;
+  const urlPageProjets = `${BASE_URL}${Routes.Projet.lister()}`;
+
+  const templateIds = {
+    dreal: 1436254,
+    partenaires: 2814281,
+  };
+
+  const templateId = match(rôle)
+    .returnType<number>()
+    .with('dreal', () => templateIds.dreal)
+    .with(
+      P.union(
+        'acheteur-obligé',
+        'grd',
+        'caisse-des-dépôts',
+        'ademe',
+        'dgec-validateur',
+        'cre',
+        'admin',
+      ),
+      () => templateIds.partenaires,
+    )
+    .exhaustive();
+
+  return [
+    {
+      templateId,
+      messageSubject: `Invitation à suivre les projets sur Potentiel`,
+      recipients: [{ email: identifiantUtilisateur }],
+      variables: {
+        invitation_link: urlPageProjets,
+      },
+    },
+  ];
+}
